items: can request api for an item
* Removes duplicate fixtures function.
* Adds sort on pickup_name location
